+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageBubbleView.swift
@@ -5,22 +5,41 @@
 import StreamChat
 import UIKit
 
-open class ChatMessageBubbleView: _View, AppearanceProvider {
-    open var roundedCorners: CACornerMask = .all {
+/// A view that displays a bubble around a message.
+public typealias ChatMessageBubbleView = _ChatMessageBubbleView<NoExtraData>
+
+/// A view that displays a bubble around a message.
+open class _ChatMessageBubbleView<ExtraData: ExtraDataTypes>: _View, AppearanceProvider, SwiftUIRepresentable {
+    /// A type describing the content of this view.
+    public struct Content {
+        /// The background color of the bubble.
+        public let backgroundColor: UIColor
+        /// The mask saying which corners should be rounded.
+        public let roundedCorners: CACornerMask
+
+        public init(backgroundColor: UIColor, roundedCorners: CACornerMask) {
+            self.backgroundColor = backgroundColor
+            self.roundedCorners = roundedCorners
+        }
+    }
+
+    /// The content this view is rendered based on.
+    open var content: Content? {
         didSet { updateContentIfNeeded() }
     }
 
     override open func setUpAppearance() {
         super.setUpAppearance()
 
-        layer.borderColor = appearance.colorPalette.border.cgColor
-        layer.cornerRadius = 16
+        layer.borderColor = appearance.colorPalette.border3.cgColor
+        layer.cornerRadius = 18
         layer.borderWidth = 1
     }
 
     override open func updateContent() {
         super.updateContent()
-
-        layer.maskedCorners = roundedCorners
+        
+        layer.maskedCorners = content?.roundedCorners ?? .all
+        backgroundColor = content?.backgroundColor ?? .clear
     }
 }