+++ Sources/StreamChatUI/CommonViews/AvatarView/ChatChannelAvatarView.swift
@@ -27,7 +27,12 @@ open class _ChatChannelAvatarView<ExtraData: ExtraDataTypes>: _View, ThemeProvid
 
     override open func updateContent() {
         guard let channel = content.channel else {
-            presenceAvatarView.avatarView.imageView.loadImage(from: nil, placeholder: appearance.images.userAvatarPlaceholder3)
+            presenceAvatarView.avatarView.imageView.loadImage(
+                from: nil,
+                placeholder: appearance.images.userAvatarPlaceholder3,
+                preferredSize: .avatarThumbnailSize,
+                components: components
+            )
             return
         }
 
@@ -49,7 +54,9 @@ open class _ChatChannelAvatarView<ExtraData: ExtraDataTypes>: _View, ThemeProvid
 
         presenceAvatarView.avatarView.imageView.loadImage(
             from: avatarURL,
-            placeholder: appearance.images.userAvatarPlaceholder4
+            placeholder: appearance.images.userAvatarPlaceholder4,
+            preferredSize: .avatarThumbnailSize,
+            components: components
         )
         presenceAvatarView.isOnlineIndicatorVisible = isOnlineIndicatorVisible
     }