+++ Sources/StreamChatUI/CommonViews/AvatarView/CurrentChatUserAvatarView.swift
@@ -72,7 +72,12 @@ open class _CurrentChatUserAvatarView<ExtraData: ExtraDataTypes>: _Control, Them
     @objc override open func updateContent() {
         let currentUserImageUrl = controller?.currentUser?.imageURL
         let placeholderImage = appearance.images.userAvatarPlaceholder1
-        avatarView.imageView.loadImage(from: currentUserImageUrl, placeholder: placeholderImage)
+        avatarView.imageView.loadImage(
+            from: currentUserImageUrl,
+            placeholder: placeholderImage,
+            preferredSize: .avatarThumbnailSize,
+            components: components
+        )
         
         alpha = state == .normal ? 1 : 0.5
     }