+++ Sources/StreamChatUI/CommonViews/QuotedChatMessageView/QuotedChatMessageView.swift
@@ -5,27 +5,40 @@
 import StreamChat
 import UIKit
 
-/// The avatar position in relation with the text message.
-struct QuotedAvatarAlignment: Equatable {
-    /// The avatar will be aligned to the left, and the message content on the right.
-    static let left = QuotedAvatarAlignment(rawValue: 0)
-    /// The avatar will be aligned to the right, and the message content on the left.
-    static let right = QuotedAvatarAlignment(rawValue: 1)
+/// The quoted author's avatar position in relation with the text message.
+/// New custom alignments can be added with extensions and by overriding the `QuotedChatMessageView.setAvatarAlignment()`.
+public struct QuotedAvatarAlignment: RawRepresentable, Equatable {
+    /// The avatar will be aligned to the leading, and the message content on the trailing.
+    public static let leading = QuotedAvatarAlignment(rawValue: 0)
+    /// The avatar will be aligned to the trailing, and the message content on the leading.
+    public static let trailing = QuotedAvatarAlignment(rawValue: 1)
 
-    private let rawValue: Int
+    public let rawValue: Int
+
+    public init(rawValue: Int) {
+        self.rawValue = rawValue
+    }
 }
 
 /// A view that displays a quoted message.
 public typealias QuotedChatMessageView = _QuotedChatMessageView<NoExtraData>
 
 /// A view that displays a quoted message.
-open class _QuotedChatMessageView<ExtraData: ExtraDataTypes>: _View, ThemeProvider {
+open class _QuotedChatMessageView<ExtraData: ExtraDataTypes>: _View, ThemeProvider, SwiftUIRepresentable {
     /// The content of the view.
     public struct Content {
         /// The quoted message.
-        let message: _ChatMessage<ExtraData>
+        public let message: _ChatMessage<ExtraData>
         /// The avatar position in relation with the text message.
-        let avatarAlignment: QuotedAvatarAlignment
+        public let avatarAlignment: QuotedAvatarAlignment
+
+        public init(
+            message: _ChatMessage<ExtraData>,
+            avatarAlignment: QuotedAvatarAlignment
+        ) {
+            self.message = message
+            self.avatarAlignment = avatarAlignment
+        }
     }
 
     /// The content of this view, composed by the quoted message and the desired avatar alignment.
@@ -35,26 +48,42 @@ open class _QuotedChatMessageView<ExtraData: ExtraDataTypes>: _View, ThemeProvid
         }
     }
 
+    /// A Boolean value that checks if all attachments are empty.
+    open var isAttachmentsEmpty: Bool {
+        guard let content = self.content else { return true }
+        return content.message.fileAttachments.isEmpty
+            && content.message.imageAttachments.isEmpty
+            && content.message.linkAttachments.isEmpty
+            && content.message.giphyAttachments.isEmpty
+    }
+
     /// The container view that holds the `authorAvatarView` and the `contentContainerView`.
-    public private(set) lazy var containerView: ContainerStackView = ContainerStackView()
+    open private(set) lazy var containerView: ContainerStackView = ContainerStackView()
         .withoutAutoresizingMaskConstraints
 
     /// The avatar view of the author's quoted message.
-    public private(set) lazy var authorAvatarView: ChatAvatarView = components
+    open private(set) lazy var authorAvatarView: ChatAvatarView = components
         .avatarView.init()
         .withoutAutoresizingMaskConstraints
 
     /// The container view that holds the `textView` and the `attachmentPreview`.
-    public private(set) lazy var contentContainerView: ContainerStackView = ContainerStackView()
+    open private(set) lazy var contentContainerView: ContainerStackView = ContainerStackView()
         .withoutAutoresizingMaskConstraints
 
     /// The `UITextView` that contains quoted message content.
-    public private(set) lazy var textView: UITextView = UITextView()
+    open private(set) lazy var textView: UITextView = UITextView()
         .withoutAutoresizingMaskConstraints
 
-    /// The attachment preview view if the quoted message has an attachment.
-    public private(set) lazy var attachmentPreviewView: UIImageView = UIImageView()
+    /// The attachments preview view if the quoted message has attachments.
+    /// The default logic is that the first attachment is displayed on the preview view.
+    open private(set) lazy var attachmentPreviewView: UIImageView = UIImageView()
         .withoutAutoresizingMaskConstraints
+
+    /// The size of the avatar view that belongs to the author of the quoted message.
+    open var authorAvatarSize: CGSize { .init(width: 24, height: 24) }
+
+    /// The size of the attachments preview.s
+    open var attachmentPreviewSize: CGSize { .init(width: 34, height: 34) }
 
     override open func setUp() {
         super.setUp()
@@ -68,10 +97,10 @@ open class _QuotedChatMessageView<ExtraData: ExtraDataTypes>: _View, ThemeProvid
 
     override open func setUpAppearance() {
         super.setUpAppearance()
+
         textView.textContainer.maximumNumberOfLines = 6
         textView.textContainer.lineBreakMode = .byTruncatingTail
         textView.textContainer.lineFragmentPadding = .zero
-
         textView.backgroundColor = .clear
         textView.font = appearance.fonts.subheadline
         textView.textContainerInset = .zero
@@ -95,21 +124,20 @@ open class _QuotedChatMessageView<ExtraData: ExtraDataTypes>: _View, ThemeProvid
         contentContainerView.isLayoutMarginsRelativeArrangement = true
         contentContainerView.alignment = .top
 
-        embed(containerView)
+        addSubview(containerView)
+        containerView.pin(to: layoutMarginsGuide)
+        directionalLayoutMargins = .zero
+
         containerView.addArrangedSubview(authorAvatarView)
         containerView.addArrangedSubview(contentContainerView)
         
         contentContainerView.addArrangedSubview(attachmentPreviewView)
         contentContainerView.addArrangedSubview(textView)
 
-        let authorAvatarSize = CGSize(width: 24, height: 24)
-
         NSLayoutConstraint.activate([
             authorAvatarView.widthAnchor.pin(equalToConstant: authorAvatarSize.width),
             authorAvatarView.heightAnchor.pin(equalToConstant: authorAvatarSize.height)
         ])
-
-        let attachmentPreviewSize = CGSize(width: 34, height: 34)
 
         NSLayoutConstraint.activate([
             attachmentPreviewView.widthAnchor.pin(equalToConstant: attachmentPreviewSize.width),
@@ -124,105 +152,93 @@ open class _QuotedChatMessageView<ExtraData: ExtraDataTypes>: _View, ThemeProvid
         guard let message = content?.message else { return }
         guard let avatarAlignment = content?.avatarAlignment else { return }
 
+        textView.text = message.text
+
         contentContainerView.backgroundColor = message.linkAttachments.isEmpty
             ? appearance.colorPalette.popoverBackground
             : appearance.colorPalette.highlightedAccentBackground1
 
         setAvatar(imageUrl: message.author.imageURL)
-        setText(message.text)
-        setAttachmentPreview(for: message)
+        setAvatarAlignment(avatarAlignment)
 
-        switch avatarAlignment {
-        case .left:
-            setAvatarPosition(.left)
-        case .right:
-            setAvatarPosition(.right)
-        default:
-            break
+        if isAttachmentsEmpty {
+            hideAttachmentPreview()
+        } else {
+            setAttachmentPreview(for: message)
+            showAttachmentPreview()
         }
     }
-}
 
-private extension _QuotedChatMessageView {
     /// Sets the avatar image from a url or sets the placeholder image if the url is `nil`.
     /// - Parameter imageUrl: The url of the image.
-    func setAvatar(imageUrl: URL?) {
+    open func setAvatar(imageUrl: URL?) {
         let placeholder = appearance.images.userAvatarPlaceholder1
-        authorAvatarView.imageView.loadImage(from: imageUrl, placeholder: placeholder)
-    }
-
-    /// Sets the text of the `textView`.
-    /// - Parameter text: The content of the text view.
-    func setText(_ text: String) {
-        textView.text = text
-    }
-
-    /// The avatar position in relation of the text bubble.
-    enum AvatarPosition {
-        case left
-        case right
+        authorAvatarView.imageView.loadImage(
+            from: imageUrl,
+            placeholder: placeholder,
+            preferredSize: .avatarThumbnailSize,
+            components: components
+        )
     }
 
     /// Sets the avatar position in relation of the text bubble.
-    /// - Parameter position: The avatar position.
-    func setAvatarPosition(_ position: AvatarPosition) {
-        authorAvatarView.removeFromSuperview()
-        switch position {
-        case .left:
+    /// - Parameter alignment: The avatar alignment of the author of the quoted message.
+    open func setAvatarAlignment(_ alignment: QuotedAvatarAlignment) {
+        containerView.removeArrangedSubview(authorAvatarView)
+
+        switch alignment {
+        case .leading:
             containerView.insertArrangedSubview(authorAvatarView, at: 0)
             contentContainerView.layer.maskedCorners = [
                 .layerMinXMinYCorner,
                 .layerMaxXMinYCorner,
                 .layerMaxXMaxYCorner
             ]
-        case .right:
+        case .trailing:
             containerView.addArrangedSubview(authorAvatarView)
             contentContainerView.layer.maskedCorners = [
                 .layerMinXMinYCorner,
                 .layerMaxXMinYCorner,
                 .layerMinXMaxYCorner
             ]
+        default:
+            break
         }
     }
 
-    /// Sets the attachment view or hides it if no attachment found in the message.
-    /// - Parameter message: The message owner of the attachment.
-    func setAttachmentPreview(for message: _ChatMessage<ExtraData>) {
+    /// Sets the attachment content to the preview view.
+    /// Override this function if you want to provide custom logic to present
+    /// the attachments preview of the message, or if you want to support your custom attachment.
+    /// - Parameter message: The message that contains all the attachments.
+    open func setAttachmentPreview(for message: _ChatMessage<ExtraData>) {
         if let filePayload = message.fileAttachments.first?.payload {
-            // TODO: Question for designers.
-            // I'm not sure if it will be possible to provide specific icon for all file formats
-            // so probably we should stick to some generic like other apps do.
             attachmentPreviewView.contentMode = .scaleAspectFit
             attachmentPreviewView.image = appearance.images.fileIcons[filePayload.file.type] ?? appearance.images.fileFallback
-            showAttachmentPreview()
+            textView.text = message.text.isEmpty ? filePayload.title : message.text
         } else if let imagePayload = message.imageAttachments.first?.payload {
             attachmentPreviewView.contentMode = .scaleAspectFill
-            attachmentPreviewView.loadImage(from: imagePayload.imageURL)
-            showAttachmentPreview()
-            // TODO: When we will have attachment examples we will set smth
-            // different for different types.
-            if message.text.isEmpty {
-                textView.text = "Photo"
-            }
+            attachmentPreviewView.loadImage(from: imagePayload.imageURL, components: components)
+            textView.text = message.text.isEmpty ? "Photo" : message.text
         } else if let linkPayload = message.linkAttachments.first?.payload {
             attachmentPreviewView.contentMode = .scaleAspectFill
-            attachmentPreviewView.loadImage(from: linkPayload.previewURL)
-            showAttachmentPreview()
-        } else {
-            attachmentPreviewView.image = nil
-            hideAttachmentPreview()
+            attachmentPreviewView.loadImage(from: linkPayload.previewURL, components: components)
+            textView.text = linkPayload.originalURL.absoluteString
+        } else if let giphyPayload = message.giphyAttachments.first?.payload {
+            attachmentPreviewView.contentMode = .scaleAspectFill
+            attachmentPreviewView.loadImage(from: giphyPayload.previewURL, components: components)
+            textView.text = message.text.isEmpty ? "Giphy" : message.text
         }
     }
 
     /// Show the attachment preview view.
-    func showAttachmentPreview() {
+    open func showAttachmentPreview() {
         Animate {
             self.attachmentPreviewView.isHidden = false
         }
     }
 
     /// Hide the attachment preview view.
-    func hideAttachmentPreview() {
+    open func hideAttachmentPreview() {
         Animate {
             self.attachmentPreviewView.isHidden = true
         }