+++ Sources/StreamChatUI/Utils/UIImageView+Extensions.swift
@@ -3,14 +3,26 @@
 //
 
 import Nuke
+import StreamChat
 import UIKit
 
 extension UIImageView {
+    /// Load image from URL
+    /// - Parameters:
+    ///   - url: URL of the image.
+    ///   - placeholder: Placeholder to show while the image is loading or in case the loading fails.
+    ///   - resize: Request thumbnail if supported and resize loaded image to match the size of the image view.
+    ///   - preferredSize: Specify the thumbnail size.
+    ///   - components: dependency injection for components.
+    ///   - completion: Image request completion block.
+    /// - Returns: Active image download task.
     @discardableResult
-    func loadImage(
+    func loadImage<ExtraData: ExtraDataTypes>(
         from url: URL?,
         placeholder: UIImage? = nil,
-        resizeAutomatically: Bool = true,
+        resize: Bool = true,
+        preferredSize: CGSize? = nil,
+        components: _Components<ExtraData>,
         completion: ImageTask.Completion? = nil
     ) -> ImageTask? {
         guard !SystemEnvironment.isTests else {
@@ -35,17 +47,23 @@ extension UIImageView {
         currentImageLoadingTask?.cancel()
 
         guard
-            let url = url
+            var url = url
         else {
             image = placeholder
             return nil
         }
 
-        let preprocessors: [ImageProcessing] = resizeAutomatically && bounds.size != .zero
-            ? [ImageProcessors.Resize(size: bounds.size, contentMode: .aspectFill, crop: true)]
+        let preprocessors: [ImageProcessing] = resize
+            ? [ImageProcessors.LateResize(sizeProvider: { [weak self] in self?.bounds.size ?? .zero })]
             : []
- 
-        let request = ImageRequest(url: url, processors: preprocessors)
+  
+        let size = preferredSize ?? bounds.size
+        if resize && size != .zero {
+            url = components.imageCDN.thumbnailURL(originalURL: url, preferredSize: size)
+        }
+        
+        let imageKey = components.imageCDN.cachingKey(forImage: url)
+        let request = ImageRequest(url: url, processors: preprocessors, options: ImageRequestOptions(filteredURL: imageKey))
         let options = ImageLoadingOptions(placeholder: placeholder)
 
         currentImageLoadingTask = Nuke.loadImage(with: request, options: options, into: self, completion: completion)
@@ -59,5 +77,43 @@ private extension UIImageView {
     var currentImageLoadingTask: ImageTask? {
         get { objc_getAssociatedObject(self, &Self.nukeLoadingTaskKey) as? ImageTask }
         set { objc_setAssociatedObject(self, &Self.nukeLoadingTaskKey, newValue, .OBJC_ASSOCIATION_RETAIN) }
+    }
+}
+
+extension ImageProcessors {
+    /// Scales an image to a specified size.
+    /// The getting of the size is offloaded via closure after the image is loaded.
+    /// The View has time to layout and provide non-zero size.
+    public struct LateResize: ImageProcessing {
+        private var size: CGSize {
+            var size: CGSize = .zero
+            DispatchQueue.main.sync { size = sizeProvider() }
+            return size
+        }
+
+        private let sizeProvider: () -> CGSize
+        
+        /// Initializes the processor with size providing closure.
+        /// - Parameter sizeProvider: Closure to obtain size after the image is loaded.
+        public init(sizeProvider: @escaping () -> CGSize) {
+            self.sizeProvider = sizeProvider
+        }
+
+        public func process(_ image: PlatformImage) -> PlatformImage? {
+            let size = self.size
+            guard size != .zero else { return image }
+            
+            return ImageProcessors.Resize(
+                size: size,
+                unit: .points,
+                contentMode: .aspectFill,
+                crop: true,
+                upscale: false
+            ).process(image)
+        }
+
+        public var identifier: String {
+            "com.github.kean/nuke/lateResize"
+        }
     }
 }